Замена на double

+++ examples/BOOST_ODE/boost_bench_sodu.cpp
@@ -3,9 +3,9 @@
 #include <iostream>
 #include <simfor/odu.hpp>
 
-float koef_matrix_settings ( unsigned j, unsigned i, unsigned n )
+double koef_matrix_settings ( unsigned j, unsigned i, unsigned n )
     {
-    float value;
+    double value;
     if ( j==i ) value = -4.f;
     else if ( j>0 && ( ( j-1 ) == i ) ) value = 1.f;
     else if ( j< ( n-1 ) && ( ( j+1 ) == i ) ) value = 1.f;
@@ -22,15 +22,15 @@ simfor::matr odu_matrix_create ( unsigned n )
     return M;
     }
 
-float integral_function ( float x, int number )
+double integral_function ( double x, int number )
     {
     return exp ( -1 * powf ( x - number, 2 ) );
     }
 
-float integral_result ( float x, int number )
+double integral_result ( double x, int number )
     {
-    float integral = 0;
-    float gauss_x[7] = {-1.f,
+    double integral = 0;
+    double gauss_x[7] = {-1.f,
                         - sqrtf ( 5. / 11. + 2. / 11. * sqrtf ( 5. / 3. ) ),
                         - sqrtf ( 5. / 11. - 2. / 11. * sqrtf ( 5. / 3. ) ), 0.0,
                         sqrtf ( 5. / 11. - 2. / 11. * sqrtf ( 5. / 3. ) ),
@@ -53,7 +53,7 @@ float integral_result ( float x, int number )
     return integral;
     }
 
-simfor::vec initial_cond ( float a, int n )
+simfor::vec initial_cond ( double a, int n )
     {
     simfor::vec y0 ( n );
     for ( unsigned i = 0; i < n; ++i )
@@ -64,7 +64,7 @@ simfor::vec initial_cond ( float a, int n )
 int main ( int argc, char* argv [] )
     {
     int n = atoi ( argv [ 1 ] );
-    float a = 0, b = 1, h;
+    double a = 0, b = 1, h;
     h = ( b - a ) / n;
     simfor::matr F = odu_matrix_create ( n ), Y;
     simfor::vec x0 = initial_cond ( 1, n );
