+++ examples/ODE/multMatrixVector.hpp
@@ -25,10 +25,10 @@ vector<C1> prod ( const matrix<C1> &R, const vector<C2> &l )
     }
 
 // template<class E1, class E2>
-// vector<float> multMatrVec_omp ( const E1 &A, const E2 &b )
+// vector<double> multMatrVec_omp ( const E1 &A, const E2 &b )
 //     {
 //     int n = A.size1() == b.size() ? b.size() : 0;
-//     vector<float> v ( n, 0 );
+//     vector<double> v ( n, 0 );
 //     #pragma omp parallel for collapse(2) schedule(auto)
 //     for ( int i = 0; i < n; i++ )
 //         {
@@ -57,7 +57,7 @@ vector<E1> prod_omp ( const matrix<E1> &A, const vector<E2> &b )
 
 
 template<class E1, class E2>
-void prod_mpi ( const matrix<E1> &A, const vector<E2> &b, vector<float> &c )
+void prod_mpi ( const matrix<E1> &A, const vector<E2> &b, vector<double> &c )
     {
 
     int n = A.size2() == b.size() ? b.size() : 0;
@@ -83,7 +83,7 @@ void prod_mpi ( const matrix<E1> &A, const vector<E2> &b, vector<float> &c )
         {
         world.Send ( &from, 1, MPI::INT, 0, 1);
         world.Send ( &to, 1, MPI::INT, 0, 2 );
-        world.Send ( &c ( from ), ( to - from ), MPI::FLOAT, 0, 3);
+        world.Send ( &c ( from ), ( to - from ), MPI::double, 0, 3);
         }
     else
         {
@@ -91,14 +91,14 @@ void prod_mpi ( const matrix<E1> &A, const vector<E2> &b, vector<float> &c )
             {
             world.Recv ( &from, 1, MPI::INT, i, 1, status );
             world.Recv ( &to, 1, MPI::INT, i, 2, status );
-            world.Recv ( &c ( from ), ( to - from ), MPI::FLOAT, i, 3, status );
+            world.Recv ( &c ( from ), ( to - from ), MPI::double, i, 3, status );
             }
         }
     MPI::Finalize();
     }
 
 template<class E1, class E2>
-vector<float> prod_mpi ( const matrix<E1> &A, const vector<E2> &b )
+vector<double> prod_mpi ( const matrix<E1> &A, const vector<E2> &b )
     {
 
     int n = A.size2() == b.size() ? b.size() : 0;
@@ -112,7 +112,7 @@ vector<float> prod_mpi ( const matrix<E1> &A, const vector<E2> &b )
     int cnt = n / p;
     int from = r * cnt;
     int to = n;
-    vector<float> c ( n );
+    vector<double> c ( n );
 
     if ( r != p-1 ) to = from + cnt;
 
@@ -125,7 +125,7 @@ vector<float> prod_mpi ( const matrix<E1> &A, const vector<E2> &b )
         {
         world.Send ( &from, 1, MPI::INT, 0, 1);
         world.Send ( &to, 1, MPI::INT, 0, 2 );
-        world.Send ( &c ( from ), ( to - from ), MPI::FLOAT, 0, 3);
+        world.Send ( &c ( from ), ( to - from ), MPI::double, 0, 3);
         }
     else
         {
@@ -133,7 +133,7 @@ vector<float> prod_mpi ( const matrix<E1> &A, const vector<E2> &b )
             {
             world.Recv ( &from, 1, MPI::INT, i, 1, status );
             world.Recv ( &to, 1, MPI::INT, i, 2, status );
-            world.Recv ( &c ( from ), ( to - from ), MPI::FLOAT, i, 3, status );
+            world.Recv ( &c ( from ), ( to - from ), MPI::double, i, 3, status );
             }
         }
     //MPI::Finalize();
